+++ src/packages/universalLoginPage/constants.ts
@@ -11,3 +11,15 @@ export const CONNECTOR_LOGO_BASE_URL =
 export const BASE_LOGO_URL = `https://cdn.jsdelivr.net/gh/sesam-io/wix-sso@${SITE_LOGOS_RELEASE_VERSION}/src/packages/siteLogos`;
 
 export const DEFAULT_LOGO_URL = `${BASE_LOGO_URL}/sesam/sesam-talk-rgb.png`;
+
+export const SiteIds = {
+  hubspot: "hubspot",
+  powerofficego: "powerofficego",
+  "powerofficego-test": "powerofficego-test",
+  sesam: "sesam",
+  superoffice: "superoffice",
+  "superoffice-test": "superoffice-test",
+  tripletex: "tripletex",
+  "tripletex-test": "tripletex-test",
+  wave: "wave",
+} as const;
