--- src/packages/universalLoginPage/siteIds.ts
@@ -1,13 +1,0 @@
-export const SiteIds = {
-  hubspot: "hubspot",
-  powerofficego: "powerofficego",
-  "powerofficego-test": "powerofficego-test",
-  sesam: "sesam",
-  superoffice: "superoffice",
-  "superoffice-test": "superoffice-test",
-  tripletex: "tripletex",
-  "tripletex-test": "tripletex-test",
-  wave: "wave",
-} as const;
-
-export const BrandedSiteIds = ["superoffice", "superoffice-test", "wave"];
